Feat: 위치 정보를 필수로 받게 로직 수정, loadin/error 컴포넌트 추가
