+++ src/app/layout.tsx
@@ -16,7 +16,6 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body>
-        <CalcGeolocation />
         <CommonLayout>{children}</CommonLayout>
       </body>
     </html>
